backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.19.0
httpx>=0.24.1
openai>=0.27.8,<0.28.0
python-multipart>=0.0.6
pydantic>=2.4.2,<3.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
langchain>=0.0.267
langchain-community>=0.0.1
faiss-cpu>=1.7.4